commit and verify database

+++ backend/src/main/java/com/intelligent/missingperson/repository/CarePartnerRepository.java
@@ -1,9 +1,10 @@
 package com.intelligent.missingperson.repository;
 
 import com.intelligent.missingperson.entity.CarePartner;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CarePartnerRepository extends JpaRepository<CarePartner, Long> {
+public interface CarePartnerRepository extends JpaRepository<CarePartner, Integer> {
 }
